project-hackathon/project/requirements.txt
scrapy==2.11.0
tabulate==0.9.0
python-dateutil==2.8.2